+++ bee-ledger/src/workers/consensus/worker.rs
@@ -24,8 +24,8 @@ use bee_runtime::{event::Bus, node::Node, shutdown_stream::ShutdownStream, worke
 use bee_tangle::{ConflictReason, MsTangle, TangleWorker};
 
 use async_trait::async_trait;
-
 use futures::{channel::oneshot, stream::StreamExt};
+use log::debug;
 use log::{error, info, warn};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::UnboundedReceiverStream;